+++ scheduler/models/__init__.py
@@ -1,3 +1,3 @@
-from .args import JobKwarg, JobArg, BaseJobArg  # noqa: F401
+from .args import TaskKwarg, TaskArg, BaseJobArg  # noqa: F401
 from .queue import Queue  # noqa: F401
-from .scheduled_job import BaseJob, ScheduledJob, RepeatableJob, CronJob  # noqa: F401
+from .scheduled_task import BaseTask, ScheduledTask, RepeatableTask, CronTask  # noqa: F401